+++ src/main/java/org/smoothbuild/lang/base/define/ImplementedBy.java
@@ -1,0 +1,4 @@
+package org.smoothbuild.lang.base.define;
+
+public record ImplementedBy(String path, boolean isPure) {
+}