+++ src/main/java/org/smoothbuild/lang/base/define/NativeBody.java
@@ -1,4 +1,4 @@
 package org.smoothbuild.lang.base.define;
 
-public record NativeBody(String implementedBy) implements Body {
+public record NativeBody(ImplementedBy implementedBy) implements Body {
 }