+++ src/main/java/org/smoothbuild/lang/parse/ast/FuncNode.java
@@ -3,11 +3,12 @@ package org.smoothbuild.lang.parse.ast;
 import java.util.List;
 import java.util.Optional;
 
+import org.smoothbuild.lang.base.define.ImplementedBy;
 import org.smoothbuild.lang.base.define.Location;
 
 public class FuncNode extends CallableNode {
   public FuncNode(Optional<TypeNode> type, String name, List<ItemNode> params,
-      Optional<ExprNode> expr, Optional<String> implementedBy, Location location) {
+      Optional<ExprNode> expr, Optional<ImplementedBy> implementedBy, Location location) {
     super(type, name, expr, params, implementedBy, location);
   }
 }