+++ src/main/test/org/smoothbuild/exec/compute/ComputationHashTest.java
@@ -77,18 +77,18 @@ public class ComputationHashTest extends TestingContext {
 
   @Test
   public void hash_of_computation_with_native_call_algorithm_and_empty_input_is_stable() {
-    Algorithm algorithm = new CallNativeAlgorithm(stringSpec(), "name", mockNative());
+    Algorithm algorithm = new CallNativeAlgorithm(stringSpec(), "name", mockNative(), true);
     Input input = input(list());
     assertThat(computationHash(Hash.of(13), algorithm, input))
-        .isEqualTo(Hash.decode("d66663fdde3d70dda784d5b035578cbd7bd6c184"));
+        .isEqualTo(Hash.decode("48cacf152c032cb117fc835fef8024660f281e4a"));
   }
 
   @Test
   public void hash_of_computation_with_native_call_algorithm_and_non_empty_input_is_stable() {
-    Algorithm algorithm = new CallNativeAlgorithm(stringSpec(), "name", mockNative());
+    Algorithm algorithm = new CallNativeAlgorithm(stringSpec(), "name", mockNative(), true);
     Input input = input(list(string("abc"), string("def")));
     assertThat(computationHash(Hash.of(13), algorithm, input))
-        .isEqualTo(Hash.decode("7343a909cf87b4465ffdd09a0ab48d6ddc9b1757"));
+        .isEqualTo(Hash.decode("816fee20334a12c3e827179a7c03701c287688aa"));
   }
 
   @Test